[WEB-3828] - renamed folder dto -> schemas
