+++ src/ecanvasser/ecanvasser.controller.ts
@@ -12,18 +12,18 @@ import {
 } from '@nestjs/common'
 import { EcanvasserService } from './services/ecanvasser.service'
 import { SurveyService } from './services/survey.service'
-import { CreateEcanvasserSchema } from './dto/createEcanvasser.schema'
-import { UpdateEcanvasserSchema } from './dto/updateEcanvasser.schema'
+import { CreateEcanvasserSchema } from './schemas/createEcanvasser.schema'
+import { UpdateEcanvasserSchema } from './schemas/updateEcanvasser.schema'
 import { PublicAccess } from 'src/authentication/decorators/PublicAccess.decorator'
 import { CampaignOwnerOrAdminGuard } from 'src/campaigns/guards/CampaignOwnerOrAdmin.guard'
 import { Roles } from 'src/authentication/decorators/Roles.decorator'
 import { ReqCampaign } from 'src/campaigns/decorators/ReqCampaign.decorator'
 import { Campaign } from '@prisma/client'
 import { UseCampaign } from 'src/campaigns/decorators/UseCampaign.decorator'
-import { CreateSurveySchema } from './dto/createSurvey.schema'
-import { CreateSurveyQuestionSchema } from './dto/createSurveyQuestion.schema'
-import { UpdateSurveyQuestionSchema } from './dto/updateSurveyQuestion.schema'
-import { UpdateSurveySchema } from './dto/updateSurvey.schema'
+import { CreateSurveySchema } from './schemas/createSurvey.schema'
+import { CreateSurveyQuestionSchema } from './schemas/createSurveyQuestion.schema'
+import { UpdateSurveyQuestionSchema } from './schemas/updateSurveyQuestion.schema'
+import { UpdateSurveySchema } from './schemas/updateSurvey.schema'
 
 @Controller('ecanvasser')
 export class EcanvasserController {
